Simplify engine exports.
* Avoid relying on initializer to setup `.reopen`'s needed.
* Use `.extend()` when possible (instead of `Engine.reopen` we can
  `.extend()`)
* Remove initializer (it is not needed internally).
* Remove all but one usage of `Ember.__loader.require` (remaining usage
  is for `hasDefaultSerializer`).
