--- addon/-private/engine-ext.js
@@ -1,19 +1,0 @@
-import Ember from 'ember';
-import EngineScopedLinkComponent from './link-to-component';
-import ExternalLinkComponent from './link-to-external-component';
-import emberRequire from './ext-require';
-
-const Engine = emberRequire('ember-application/system/engine');
-
-Engine.reopen({
-  buildRegistry() {
-    let registry = this._super(...arguments);
-
-    if (!(this instanceof Ember.Application)) {
-      registry.register('component:link-to', EngineScopedLinkComponent);
-      registry.register('component:link-to-external', ExternalLinkComponent);
-    }
-
-    return registry;
-  }
-});
