+++ addon/-private/engine-instance-ext.js
@@ -1,10 +1,8 @@
 import Ember from 'ember';
-import emberRequire from './ext-require';
-
-const EngineInstance = emberRequire('ember-application/system/engine-instance');
 
 const {
-  assert
+  assert,
+  EngineInstance
 } = Ember;
 
 EngineInstance.reopen({
