+++ addon/engine.js
@@ -1,7 +1,23 @@
-import emberRequire from './-private/ext-require';
+import Ember from 'ember';
+import EngineScopedLinkComponent from './-private/link-to-component';
+import ExternalLinkComponent from './-private/link-to-external-component';
 
-// Because feature flags are only valid for Ember canary builds,
-// `Ember.Engine` won't be exposed publicly for beta builds. This export
-// provides a means to use this module with any ember build which privately
-// includes the `engine` module.
-export default emberRequire('ember-application/system/engine');
+// Load extensions to Ember
+import './-private/route-ext';
+import './-private/router-ext';
+import './-private/engine-instance-ext';
+
+const { Engine } = Ember;
+
+export default Engine.extend({
+  buildRegistry() {
+    let registry = this._super(...arguments);
+
+    if (!(this instanceof Ember.Application)) {
+      registry.register('component:link-to', EngineScopedLinkComponent);
+      registry.register('component:link-to-external', ExternalLinkComponent);
+    }
+
+    return registry;
+  }
+});
