--- addon/initializers/engines.js
@@ -1,14 +1,0 @@
-// Load extensions to Ember
-import '../-private/route-ext';
-import '../-private/router-ext';
-import '../-private/engine-ext';
-import '../-private/engine-instance-ext';
-
-// TODO: Move to ensure they run prior to instantiating Ember.Application
-export function initialize() {
-}
-
-export default {
-  name: 'engines',
-  initialize
-};
