--- app/initializers/engines.js
@@ -1,1 +1,0 @@
-export { default, initialize } from 'ember-engines/initializers/engines';
